In a nutshell, you are asked to type your current bill. From there, you would rate your service(either poor-average-excellent).
If you believe the service was poor, you would tip 10% of the bill, 15% if the service was average, and 20% of the bill if it was excellent service.

From than your bill and the service rating will both be multiplied and than you'll get your tip.

completed the regular tip calculator

Need to figure out/add:

title

whether or not im going to also add a party tip calculator(need to see how those work).

test out project before i deploy it